Based on: Disable user-select on widgets #696
This might be the worst react code ever (I've never used react before 💀). I have no other idea on how to do this. Somebody else please fix it. It does work great though, adds a setting to enable/disable highlighting of the widgets as some people might want this and others hate it. DEFAULT is TRUE