Add a modal for selecting the icon.
Right now it only support feather, which is what it was before, but when you go to selecting and icon it opens a modal with search, etc... It also is backwards-compatible!
The styles might be a bit off from the rest of the app, but that can be fixed later.